cephadm: use ceph.git's cephadm as binary for dev
Given the constant mutability of cephadm internals, we can't reliably
rely on including it as a library in the short-term [1], so we are
better off using it as a binary and calling external commands.

For development purposes, let's use whatever is pinned in the ceph.git
submodule repository. For deployments we can instead rely on whatever
has been installed in the system image.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>