glass: Allow bootstrapped cluster to access the dashboard
Currently a cluster that is not in the "ready" state can't access the
dashboard, but this state can not be achieved yet as it's not
implemented. With a bootstrapped cluster however we can already
show data on the dashboard, this is why it should be possible to see it.

Possible means you can access it if you know the URL. It won't be made
visible to the user during the installation process yet as this is TBD.

Signed-off-by: Stephan Müller <smueller@suse.com>