Introduce runtime python venv for dependencies, add Makefile
This is work towards being able to have an actual package (rpm, deb, ...)
of Aquarium (see https://github.com/aquarist-labs/aquarium/issues/91).

I've added a Makefile with a `dist` target, so you can now run `make dist`
to get a tarball named something like aquarium-0.1.0-0-gc28eb76.tar.gz,
which includes a built glass and all the necessary bits of gravel.  It
also includes all our python dependencies (per src/requirements.txt) in
/usr/share/aquarium/venv, so we don't have to worry about not having
packaged versions of these things on whatever distros we deploy on.

This doesn't change the current dev workflow at all.  You still build an
image by running `./tools/build-image.sh`, but that script will in turn
call `make dist` to get a tarball to install in the image.  In future
(once I've got a spec file and rpm build sorted out), this can be changed
so that the kiwi config uses that actual package, instead of the tarball.

That said, if you *don't* want to build an image, but would rather try
a manual Aquarium install on some random Linux, you could run `make dist`,
then take the tarball and extract it on the system you want to deploy
on.  Might be a bit of fun ;-)

Note: the version string used in the tarball name is based on the latest
annotated tag in the form 'vX.Y.Z', which means I've suddenly introduced
the notion of actual version numbers and releases to the project.  Sorry.
Until such time as an actual release is tagged, the version will be set
to v0.0.0.

Fixes: https://github.com/aquarist-labs/aquarium/issues/277
Signed-off-by: Tim Serong <tserong@suse.com>