gravel: nodes/mgr: make ceph.conf managed by cephadm
So we can have a consistent and up to date ceph.conf across all nodes,
let's rely on cephadm to manage it.

We don't drop the ceph.conf initially shared with a joining node because
we want to ensure that node is able to perform operations on the cluster
as soon as join finishes, and we don't want to have to wait for cephadm
to write the ceph.conf.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>