tools: Add Makefile for building dist tarball
I've added a Makefile with a `dist` target, so you can now run `make dist`
to get a tarball named something like aquarium-0.1.0-0-g2b6942a.tar.gz,
which includes a built glass and all the necessary bits of gravel.

This doesn't change the current dev workflow at all.  You still build an
image by running `./tools/build-image.sh`, but that script will in turn
call `make dist` to get a tarball to install in the image.

Note: the version string used in the tarball name is based on the latest
annotated tag in the form 'vX.Y.Z', which means I've suddenly introduced
the notion of version numbers and releases to the project.  Until such
time as an actual release is tagged, the version will default to v0.0.0,
just to have the build working.

Fixes: https://github.com/aquarist-labs/aquarium/issues/277
Signed-off-by: Tim Serong <tserong@suse.com>