tools: setup-dev.sh: check for python >= 3.8
This ensures we've got at least python 3.8, which is what we use
in our images, and is required for asyncio (it's not impossible
3.7 might work, but nobody's tested that so it seems safest to
stick with 3.8).

I've also made it a bit more obvious when dependency installation
fails.  I've seen cases where zypper fails to refresh some repo,
but that scrolls off the top of the screen, the script bails out
and all you see at the bottom is the "Nothing to do" line from
zypper, so you just assume there's nothing to do.

Signed-off-by: Tim Serong <tserong@suse.com>