images: config.sh: install aetcd3 direct from github
Pip knows how to install Python packages directly from git repos,
so let's take advantage of that for our tweaked aetcd3.  Note that
this *does* mean adding git to the image in order for the build
to work, and I'm not sure this is entirely desirable.  What's
everyone else think?

Signed-off-by: Tim Serong <tserong@suse.com>