gravel: adding third node now works with etcd
While our approach was working for two nodes, the fact is that for more
nodes we were missing peer information for the initial cluster. Instead,
what we should have been doing is relying on the existing nodes result
provided by aetcd3's "client.add_member()" (which requires a patched
version of aetcd3 at time of this commit, and for which support has been
included in a previous patch).

Signed-off-by: Joao Eduardo Luis <joao@suse.com>