cephfs: Do not maintain a persistent instance
CephFS is mostly a helper class for cephfs commands. There is no state
tracked in it and no need to create a persistent instance.

I'm doing this as a separate commit (as opposed to fixing the original
change) so we can revert it down the line if we decide keeping one
instance is helpful (eg for caching results).

Signed-off-by: Joshua Hesketh <josh@nitrotech.org>