tools/build-image: make sure _out contents can be read
By some reason kiwi-ng on ubuntu creates all files with 600 mode,
in combination with root ownership this makes it impossible to
add vagrant box using the images found in the directory.

I have idea how to fix this correctly in this moment, so just
adding read permission to all a posteriori.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.com>