nodes: deploy: create system disk
If a solution for disk usage is found, and it's possible to deploy,
deploy a system disk according to this solution; otherwise refuse to
deploy.

The system disk is going to be created on a specific device, and is
going to be partitioned and formatted with aquarium usage in mind. This
is where aquarium-specific data is going to be persisted. Depending on
what we intend to persist, individual directories will be created in
this partition (e.g., ceph, etcd, ...), and either overlayed on top of
existing file systems, or bind mounted to a specific location in the
root tree.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>