glass: Add missing en_US.json file
The regression was introduced by https://github.com/aquarist-labs/aquarium/pull/545.

NOTE: This file has to be included, even if it is empty (an empty dictionary). The i18n library want to load it. This is by intention to be able to translate english texts if they are spelled incorrect or need some adaptions somehow.

Signed-off-by: Volker Theile <vtheile@suse.com>