images: ensure we setup right on real server
While with Vagrant we have a persistent install disk, on a real server
we are running out of a ramdisk and we thus have no such benefit. This
means we need to start at the right time to ensure we mount the
systemdisk before certain things happen, and we need to ensure we setup
the hostname manually as well.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>