tools/libaqr/deployment: Remove 'host_ip' from node port forward
With the virtualbox provider having host_ip = "*" causes vagrant to
fail with `getaddrinfo: Name or service not known (SocketError)`, when
trying to getaddrinfo(*).

According to https://www.vagrantup.com/docs/networking/forwarded_ports
not setting the parameter is equal to binding to every IP. Removing
the setting should therefore have the same meaning.

Signed-off-by: Marcel Lauhoff <marcel.lauhoff@suse.com>