gravel: add pool and mds's remove
Add a function to remove pools from the cluster. In order to do so, we
need to ensure the config option `mon_allow_pool_delete` has been set
to `True` beforehand. Also set it back to `False` afterwards.

Add a function `rm_mds` to remove mds's. This will e.g. be needed in
order to remove configured CephFS services.
Also add `get_daemons` to check if there are still any running daemons
of a given service. E.g. to be able to wait for the remove operation to
be completed.

Signed-off-by: Tatjana Dehler <tdehler@suse.com>