glass: Widget glow
Currently only the cluster health widget can show all widget states, all
others will just show the blue info glow as their data is yet not
interpreted.

The possible colors and their meanings:

WidgetHealthStatus.info for uninterpreted data where the border color glow
is blue.

WidgetHealthStatus.success for interpreted data which is in a good shape.
The border color glow is green.

WidgetHealthStatus.warning for interpreted data which is in a warning
state and the user has to do something in near future in order to
prevent a failure. The border color glow is yellow.

WidgetHealthStatus.error for interpreted data which is in an error
state and the user has to do something now in order to resolve the
problem. The border color glow and the background of the widget is red.

Fixes #474
Signed-off-by: Stephan Müller <smueller@suse.com>