gravel: kv: make ensure_connection() do something useful
This allows us to kick the connection thread to try to actually get
a connection to the cluster when joining new nodes, so that we can
pull needed values from the kvstore (e.g. ntp_addr).  Note that if
you watch the logs, on the joining node, you'll first see:

-- kv -- ensure_connection: no cluster exists yet after 5 tries

Then, a second or so later, you'll see:

-- kv -- ensure_connection: cluster state 'connected' after 1 tries

The first invocation is coming from the call to ensure_connection()
from GlobalState.init_store().  At that point in time, there's no
cluster yet, so that's fine.  The second invocation is from
NodeDeployment.join(), once we can actually access the cluster.

Signed-off-by: Tim Serong <tserong@suse.com>