gravel: call kvstore ensure_connection() after bootstrap
This is not technically necessary (the cluster connection thread will
have well and truly found the cluster and connected to it by now).
But I thought it was neater to make this explicit in the bootstrap
case.  Also, if we ever change the sleep in the connection thread to
something signfificantly longer than 10 seconds, this might actually
become necessary.

Signed-off-by: Tim Serong <tserong@suse.com>