deps: Fix potential dependency issues on dev machines
Because setup-dev.sh is ran on the developers machine, it is not okay
to use the system packages which may conflict with the requirements.txt
aquarium is trying to install (for example, the system packages may
already supply a dependency which will mean that pip will then not
install it into the venv).

Instead we should ensure the venv is created without system packages.
This is what is eventually copied into the virtual machine. Before we
copy it in though we need to re-enable the system packages for
librados to work, so we do that manually with sed.

Signed-off-by: Joshua Hesketh <josh@nitrotech.org>