hardware: Validate individual node requiements
This is a very crude validation of the localhosts hardware requirements.
The frontend can check that the node passes validation before allowing
it to be bootstrapped or enrolled into the cluster.

No ongoing monitoring occurs, and we could still allow a user to proceed
to bootstrap if they wanted to.

Signed-off-by: Joshua Hesketh <jhesketh@suse.com>