images: add comment about man pages not being installed
When the image is built, <rpm-excludedocs>true</rpm-excludedocs>
makes sure man pages aren't installed, which is one of those things
you exclude to try to keep image sizes down.  This is the build-time
equivalent of rpm.install.excludedocs in /etc/zypp/zypp.conf.

I'm not suggesting we change this setting, just documenting it so
that the next time someone's poking around and wondering why there's
no man pages, we have the answer.

Signed-off-by: Tim Serong <tserong@suse.com>