images/aquarium: allow ssh-rsa algorithms in vagrant images
OpenSSH 8.8 (a fairly recent update in Tumbleweed) disables RSA
signatures using the SHA-1 hash algorithm by default for security
reasons (see https://www.openssh.com/txt/release-8.8).

This is problematic for older versions of ruby's Net::SSH library
included with openSUSE Leap 15.2, which don't yet understand the new
rsa-sha2-512 and rsa-sha2-25 host key algorithms.  This means you
can't actually use Aquarium's Tumbleweed-based vagrant images on a
Leap 15.2 host, because vagrant on the host can't ssh to the guests.
It just spins forever on "Waiting for SSH to become available...".

Thankfully we can work around this by configuring sshd in the
Vagrant images to allow the crusty old ssh-rsa algorithms.  IMO
this doesn't constitute a security risk, because we're only doing
it for vagrant images, which are used for development and testing
purposes, not for production clusters.

Signed-off-by: Tim Serong <tserong@suse.com>