tools/build-image: Use kiwi's "--logfile stdout" to catch debug output
kiwi >= 9.23.61 will try to run scripts inside `screen` if --debug is
specified.  That's fine when run from a terminal, but breaks the CI when
run inside Jenkins.  The latest versions of kiwi (>= 9.23.63) have a
"--logfile stdout" option to dump debug output to stdout, without trying
to run scripts inside a screen session, so let's use that if it's
available and there's no terminal active.

Signed-off-by: Tim Serong <tserong@suse.com>