gravel: add network controller and API endpoints
This adds the ability for gravel to read and write network interface
configuration, nameservers and routes.  We've added an
/api/local/network/config endpoint.  GET this, and you'll retrieve
a dict of interfaces, a list of nameservers and a list of static
routes.  In the simple case for a single NIC that's configured for
DHCP, this'll give you:

{
  "interfaces": {
    "eth0": {
      "name": "eth0",
      "config": {
        "bootproto": "dhcp",
        "ipaddr": "",
        "bonding_slaves": []
      }
    }
  },
  "nameservers": [],
  "routes": []
}

POST back to that same URL to apply a different config.  The config
exposed is deliberately a subset of what's possible to specify in
/etc/sysconfig/network/{config,routes,ifcfg-*,ifroute-*} files.  For
'interfaces', bootproto can be either "dhcp" or "static".  If it's
"static", you also need to specify "ipaddr", and that needs to include
the netmask or equivalent on the end (e.g.: "192.168.121.1/24").
If you want to create a bond, add a new interface to that dict and
set bonding_slaves to a list of the physical interfaces to bond.  If
the list is empty, it's just a normal interface.

'nameservers' is just a list of IP addresses, e.g.: ["8.8.8.8", ...]

'routes' is a list of dicts, each of which needs to include
'destination', 'gateway' and 'interface' as described in `man ifroute`.

When you apply the config, all network services are restarted.  Be
particularly careful when applying a new config to the interface via
which you're accessing the API.  Switch from DHCP to static IP and
forget to configure DNS and/or a default route, and you can get yourself
in trouble.  Likewise if you change the IP address to be on some other
network.

Note that for NICs we've got an InterfaceModel which includes an
InterfaceConfigModel.  This was based on Joao's work in
https://github.com/aquarist-labs/aquarium/pull/603, the idea being
that we may also want to add stats and other data to InterfaceModel.
OTOH, that may duplicate information we already get from cephadm,
so we might want to consider collapsing this instead.

Signed-off-by: Tim Serong <tserong@suse.com>